backend/requirements.txt
fastapi[standard]==0.128.2
uvicorn==0.40.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
alembic==1.18.3
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.22
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
flake8==7.3.0